Get vsphere Host; Separate graph and graphql
