+++ pkg/controller/provider/web/doc.go
@@ -4,7 +4,7 @@ import (
 	"github.com/konveyor/controller/pkg/inventory/container"
 	libweb "github.com/konveyor/controller/pkg/inventory/web"
 	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/base"
-	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/graph"
+	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/graphql"
 	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/ocp"
 	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/ovirt"
 	"github.com/konveyor/forklift-controller/pkg/controller/provider/web/vsphere"
@@ -32,6 +32,6 @@ func All(container *container.Container) (all []libweb.RequestHandler) {
 		ovirt.Handlers(container)...)
 	all = append(
 		all,
-		graph.Handlers(container)...)
+		graphql.Handlers(container)...)
 	return
 }
