+++ pkg/controller/provider/web/graphql/doc.go
@@ -1,4 +1,4 @@
-package graph
+package graphql
 
 import (
 	"github.com/konveyor/controller/pkg/inventory/container"
